add `ErrOrderFilled`
