+++ pkg/errs/errors.go
@@ -3,3 +3,7 @@ package errs
 import "errors"
 
 var OrderNotFound = errors.New("order not found")
+
+var (
+	ErrOrderFilled = errors.New("Order has been filled")
+)
